license: AGPL-3.0 -> MIT (#32)
In this pull request, we are transitioning the repository's license from AGPL v3 to MIT. This adjustment is in response to numerous user requests for a more permissive license. The AGPL's strict requirement for both modifications and any frontend using the API to be fully open-sourced under AGPL would potentially restrict the software's user base. Given that the software is intended for use as a deployment server, this change aims to enhance its accessibility.

All contributors are included as reviewers in this request. Kindly approve these modifications if you prefer your contributions to be published under the MIT license instead of the original AGPL.